requirements.txt
streamlit
Pillow
pandas
google-generativeai
python-dotenv
